Giai đoạn 1: Cà phê nhân xanh, Rang/Sản xuất, Kiểm kê, Dashboard kho
Mở rộng module Kho theo yêu cầu quản lý kinh doanh đầy đủ — audit toàn bộ
schema/API/UI hiện có trước khi code, KHÔNG tạo bảng trùng, tái sử dụng
tối đa (inventory, inventory_transactions, adjustInventory có sẵn).

THÊM MỚI:
- Nhóm "Cà phê nhân xanh" / "Cà phê rang rời" — thêm cột products.coffee_stage
  (nullable, ADD COLUMN, KHÔNG rename/recreate products vì bảng này có
  product_variants + devices là con — đã kiểm chứng thực nghiệm việc rename
  sẽ làm hỏng FK của 2 bảng đó, rủi ro không chấp nhận được cho dữ liệu
  production). Cà phê đóng gói hiện tại không đổi (coffee_stage NULL).
- Cho phép số lượng thập phân (vd 3.5kg) khi nhập/điều chỉnh kho — CHỈ cho
  nhóm nhân xanh/rang rời, các SKU khác vẫn bắt buộc số nguyên như cũ.
- Module Rang/Sản xuất: roast_batches (DRAFT/CONFIRMED) + roast_cost_config
  — hao hụt CHỈ áp dụng ở bước rang (không áp dụng lúc nhập/bán), công thức
  giá vốn/kg đã khớp chính xác ví dụ trong yêu cầu (139.625đ/kg với giá
  nhân xanh 106.000đ, hao hụt 20%, gas 1.700đ, bao bì 5.000đ). Xác nhận mẻ
  rang chỉ trừ/cộng tồn đúng 1 lần (unique index + compare-and-set, giống
  cơ chế issueInventoryForOrder có sẵn).
- Kiểm kê: trang nhập tồn thực tế, tự tính chênh lệch, xác nhận qua
  adjustInventory có sẵn (không sửa trực tiếp số tồn, bắt buộc lý do).
- Lịch sử giao dịch kho: xem/lọc inventory_transactions theo loại.
- Dashboard Kho (Tổng quan): giá trị tồn kho, tổng SKU, tồn nhân xanh/rang
  tách biệt (không trộn khi cảnh báo), tổng máy/thiết bị, sản lượng rang
  khả dụng từ nhân xanh hiện có.

SỬA:
- inventory_transactions.type: thêm ROAST_PRODUCTION/ROAST_CONSUMPTION
  (rename+recreate an toàn — đã xác nhận không có bảng nào FK tới bảng này).
- Trang Kho cũ (bảng tồn + máy/serial) giữ nguyên logic, chuyển sang
  /inventory/stock; /inventory giờ là Tổng quan mới.

TEST: 12 test case mới (Test Case 1/2/3/12 trong yêu cầu + double-submit
protection cho mẻ rang + chặn số lẻ cho SKU thường + chặn rang khi không
đủ nhân xanh) — toàn bộ 64 test pass. Đã kiểm thử thủ công đầy đủ luồng
UI thật qua Playwright: tạo SKU nhân xanh/rang → nhập 500kg → tạo mẻ rang
100kg → xác nhận → tồn kho đúng 400kg/80kg → kiểm kê → dashboard đúng số.

CHƯA LÀM (để giai đoạn sau, xem báo cáo chi tiết gửi kèm): báo giá, VAT
cấu hình theo %, bậc giá sỉ theo tổng KG, danh bạ nhà cung cấp, báo cáo
kho/lợi nhuận đầy đủ, Google Sheets backup, phiếu nhập kho dạng DRAFT/
CONFIRMED cho hàng hóa thường (nhập kho hiện tại vẫn 1-bước nhưng đã đúng
100% nguyên tắc "nhập đúng số lượng thực tế, không tự trừ hao hụt").

+++ crm/src/app/api/inventory/adjust/route.ts
@@ -7,7 +7,9 @@ import { handleApiError, ValidationError } from "@/lib/api/errors";
 
 const bodySchema = z.object({
   productVariantId: z.string().min(1),
-  delta: z.number().int(),
+  // Số nguyên hay thập phân tùy loại SKU — kiểm tra thực sự nằm ở
+  // adjustInventory() (chỉ nhóm tồn theo KG lẻ mới được số lẻ).
+  delta: z.number(),
   note: z.string().trim().min(1, "Vui lòng nhập lý do điều chỉnh"),
 });
